fix(models): change RPM property type from int to decimal
- Update RPM type to decimal in ProductionAllotment model and related DTOs
- Adjust database model snapshot for RPM column with appropriate precision
- Modify appsettings with new connection string and printer IP addresses

+++ DTOs/StorageCapture/StorageCaptureDTOs.cs
@@ -122,7 +122,7 @@ namespace AvyyanBackend.DTOs.StorageCapture
 		public int MachineId { get; set; }
 		public int NumberOfNeedles { get; set; }
 		public int Feeders { get; set; }
-		public int RPM { get; set; }
+		public decimal RPM { get; set; }
 		public decimal RollPerKg { get; set; }
 		public decimal TotalLoadWeight { get; set; }
 		public decimal TotalRolls { get; set; }
